Fixed bugs and now builds active diff cpp refac

+++ firmware/hexray/VC/src/app/torque_vectoring/controllers/yaw_rate_control/app_activeDifferential.cpp
@@ -1,12 +1,12 @@
-#include "app_activeDifferential.h"
-#include "app_vehicleDynamicsConstants.h"
+#include "app_activeDifferential.hpp"
+#include "app_vd_constants.hpp"
 #include "app_units.h"
 
 #include <cmath>
 
 // Maximum motor speed limit for clamping (6000 RPM converted to rad/s)
-constexpr double MAX_MOTOR_SPEED_RPM   = 6000.0;
-constexpr double MAX_MOTOR_SPEED_RAD_S = RPM_TO_RADS(MAX_MOTOR_SPEED_RPM);
+constexpr float MAX_MOTOR_SPEED_RPM   = 6000.0f;
+constexpr float MAX_MOTOR_SPEED_RAD_S = RPM_TO_RADS(MAX_MOTOR_SPEED_RPM);
 
 /**
  * Converts steering wheel angle to wheel angle using approximation constant.
@@ -15,11 +15,12 @@ constexpr double MAX_MOTOR_SPEED_RAD_S = RPM_TO_RADS(MAX_MOTOR_SPEED_RPM);
  * @param steering_wheel_angle_deg Steering wheel angle in degrees
  * @return Wheel angle in radians
  */
-static double steeringToWheelAngle(const double steering_wheel_angle_deg)
+static float steeringToWheelAngle(const float steering_wheel_angle_deg)
 {
     // Placeholder function - will be replaced with proper transfer function
     // Currently uses approximation constant APPROX_STEERING_TO_WHEEL_ANGLE
-    return DEG_TO_RAD(steering_wheel_angle_deg * APPROX_STEERING_TO_WHEEL_ANGLE);
+    return static_cast<float>(
+        DEG_TO_RAD(steering_wheel_angle_deg * static_cast<float>(APPROX_STEERING_TO_WHEEL_ANGLE)));
 }
 
 extern "C"
@@ -52,29 +53,29 @@ extern "C"
      * @return Motor speed references in rad/s (inner and outer wheels allocated appropriately for 4WD)
      */
     TorqueAllocationOutputs app_activeDifferential_computeTorque(
-        const double omega_v_ref,
-        const double steering_wheel_angle_deg,
-        const double omega_m_max)
+        const float omega_v_ref,
+        const float steering_wheel_angle_deg,
+        const float omega_m_max)
     {
         // Convert steering wheel angle to wheel angle (delta)
-        const double delta = steeringToWheelAngle(steering_wheel_angle_deg);
+        const float delta = steeringToWheelAngle(steering_wheel_angle_deg);
 
         // Step 1: Compute wheel speed difference Δω (Eq. 13 from paper)
         // Δω = (TRACK_WIDTH_m/WHEELBASE_m) * tan(δ) * ω_v
-        const double delta_omega = (TRACK_WIDTH_m / WHEELBASE_m) * std::tan(delta) * omega_v_ref;
+        const float delta_omega = (TRACK_WIDTH_m / WHEELBASE_m) * tanf(delta) * omega_v_ref;
 
         // Step 2: Split base speed into left/right wheel references (Eq. 15-16 from paper)
         // Note: The sign of delta_omega determines which side is inner/outer
-        const double omega_L_wheel_ref = omega_v_ref + 0.5 * delta_omega;
-        const double omega_R_wheel_ref = omega_v_ref - 0.5 * delta_omega;
+        const float omega_L_wheel_ref = omega_v_ref + 0.5f * delta_omega;
+        const float omega_R_wheel_ref = omega_v_ref - 0.5f * delta_omega;
 
         // Step 3: Convert wheel refs to motor refs using gear ratio (Eq. 17-18 from paper)
-        double omega_L_motor_ref = GEAR_RATIO * omega_L_wheel_ref;
-        double omega_R_motor_ref = GEAR_RATIO * omega_R_wheel_ref;
+        float omega_L_motor_ref = GEAR_RATIO * omega_L_wheel_ref;
+        float omega_R_motor_ref = GEAR_RATIO * omega_R_wheel_ref;
 
         // Step 4: Limit motor speeds for safety (clamp to prevent exceeding 6000 RPM limit)
         // Use provided omega_m_max if > 0, otherwise use default MAX_MOTOR_SPEED_RAD_S
-        const double max_speed_limit = (omega_m_max > 0.0) ? omega_m_max : MAX_MOTOR_SPEED_RAD_S;
+        const float max_speed_limit = (omega_m_max > 0.0f) ? omega_m_max : MAX_MOTOR_SPEED_RAD_S;
 
         // Clamp motor speeds to maximum allowed value
         if (omega_L_motor_ref > max_speed_limit)
@@ -107,9 +108,9 @@ extern "C"
         // - delta == 0 (straight): delta_omega == 0, so omega_L == omega_R
         //   -> All wheels same speed
 
-        double omega_fl, omega_fr, omega_rl, omega_rr;
+        float omega_fl, omega_fr, omega_rl, omega_rr;
 
-        if (delta > 0.0)
+        if (delta > 0.0f)
         {
             // Right turn: right wheels (FR, RR) are inner (slower), left wheels (FL, RL) are outer (faster)
             omega_fl = omega_L_motor_ref; // Front left - outer (faster)
@@ -117,7 +118,7 @@ extern "C"
             omega_rl = omega_L_motor_ref; // Rear left - outer (faster)
             omega_rr = omega_R_motor_ref; // Rear right - inner (slower)
         }
-        else if (delta < 0.0)
+        else if (delta < 0.0f)
         {
             // Left turn: left wheels (FL, RL) are inner (slower), right wheels (FR, RR) are outer (faster)
             omega_fl = omega_L_motor_ref; // Front left - inner (slower)
@@ -137,10 +138,10 @@ extern "C"
         // Return motor speeds allocated to inner/outer wheels
         // note: Currently returning in TorqueAllocationOutputs structure (using quintuna structure format)
         TorqueAllocationOutputs result;
-        result.front_left_torque  = static_cast<float>(omega_fl); // Front left motor speed (rad/s)
-        result.front_right_torque = static_cast<float>(omega_fr); // Front right motor speed (rad/s)
-        result.rear_left_torque   = static_cast<float>(omega_rl); // Rear left motor speed (rad/s)
-        result.rear_right_torque  = static_cast<float>(omega_rr); // Rear right motor speed (rad/s)
+        result.front_left_torque  = omega_fl; // Front left motor speed (rad/s)
+        result.front_right_torque = omega_fr; // Front right motor speed (rad/s)
+        result.rear_left_torque   = omega_rl; // Rear left motor speed (rad/s)
+        result.rear_right_torque  = omega_rr; // Rear right motor speed (rad/s)
         return result;
     }
 } // extern "C"
